+++ src/Platform/Http/Controllers/Posts/PostController.php
@@ -36,7 +36,7 @@ class PostController extends Controller
      */
     public function index(PostBehavior $type) : View
     {
-        $this->checkPermission('dashboard.posts.' . $type->slug);
+        $this->checkPermission('dashboard.posts.type.' . $type->slug);
 
         return view('dashboard::container.posts.main', $type->generateGrid());
     }
@@ -48,7 +48,7 @@ class PostController extends Controller
      */
     public function create(PostBehavior $type) : View
     {
-        $this->checkPermission('dashboard.posts.' . $type->slug);
+        $this->checkPermission('dashboard.posts.type.' . $type->slug);
 
         $locales = (method_exists($type, 'locale')) ? collect($type->locale()) : $this->locales;
 
@@ -67,7 +67,7 @@ class PostController extends Controller
      */
     public function store(Request $request, PostBehavior $type, Post $post) : RedirectResponse
     {
-        $this->checkPermission('dashboard.posts.' . $type->slug);
+        $this->checkPermission('dashboard.posts.type.' . $type->slug);
         $this->validate($request, $type->rules());
 
         $post->fill($request->all());
@@ -84,7 +84,7 @@ class PostController extends Controller
             $slug = $request->get('slug');
         } else {
             $content = $request->get('content');
-            $slug = reset($content)[$type->slugFields];
+            $slug = $type->slugFields ? reset($content)[$type->slugFields] : 1;
         }
 
         $post->slug = SlugService::createSlug(Post::class, 'slug', $slug);
@@ -114,7 +114,7 @@ class PostController extends Controller
      */
     public function edit(PostBehavior $type, Post $post) : View
     {
-        $this->checkPermission('dashboard.posts.' . $type->slug);
+        $this->checkPermission('dashboard.posts.type.' . $type->slug);
 
         $locales = (method_exists($type, 'locale')) ? collect($type->locale()) : $this->locales;
 
@@ -131,11 +131,10 @@ class PostController extends Controller
      * @param Post         $post
      *
      * @return \Illuminate\Http\RedirectResponse
-     * @throws \Orchid\Platform\Exceptions\TypeException
      */
     public function update(Request $request, PostBehavior $type, Post $post) : RedirectResponse
     {
-        $this->checkPermission('dashboard.posts.' . $type->slug);
+        $this->checkPermission('dashboard.posts.type.' . $type->slug);
         $post->fill($request->except('slug'));
         $post->user_id = Auth::user()->id;
 
@@ -187,7 +186,7 @@ class PostController extends Controller
      */
     public function destroy(PostBehavior $type, Post $post) : RedirectResponse
     {
-        $this->checkPermission('dashboard.posts.' . $type->slug);
+        $this->checkPermission('dashboard.posts.type.' . $type->slug);
         try {
             $post->delete();
         } catch (\Exception $e) {