+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+## [2.0.9] - 2017-12-29
+### Changed
+- Optimization for parser
+
+### Fixed
+- Bug name permission for pages
+
+## [2.0.8] - 2017-12-27
 ### Added 
 - Input Mask
 
@@ -13,8 +22,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Section wrapper remove
 - Input build named
 - "Not found" message occupies the entire screen
+- Error 403 to 404
 
 ### Fixed
+- Bug permission for behaviors
 - Style for select2
 - Calling relation for new object in template #382